// src/Application/DTOs/Auth/Auth.dto.ts
import type { KYCStatus, Role } from "../../../Infrastructure/Types/types";

export interface SendOtpDTO {
  name: string;
  email: string;
  mobile?: string;
  password: string;
  role: Role;
  isResend?: boolean;
};

export interface OtpResponseDTO {
  email: string;
  role: string;
  expiresAt: number;
};

export interface VerifyOtpResponseDTO {
  success: boolean;
  message: string;
  user: {
    id: string;
    name: string;
    email: string;
    role: Role;
    kycStatus:KYCStatus;
    accessToken: string
  }
};

export type SendLoginResponse = VerifyOtpResponseDTO;

export interface UserDTO {
  name: string;
  email: string;
  mobile?: string | null;
  password?: string | null;
  role: Role;
}



export interface LoginDTO {
  email: string;
  password: string;
  role: Role;
}

export interface LogoutDTO {
  id: string
  role: Role;
}

export interface SendLogoutResponseDTO {
  success: boolean;
  message: string;
}


export interface ResendOtpDTO {
  email: string;
  role: string;
}

export interface VerifyOtpDTO {
  email: string;
  otp: string;
  role: string;
}

export interface ForgotPasswordDTO {
    email: string;
    role:Role
}

export interface ResetPasswordDTO {
  password: string;
  role: Role;
}

export interface GoogleUserDTO {

    googleId: string;

    name: string;

    email: string;

    authProvider: "google";

    emailVerified: boolean;
}